refactor: renaming StatWidget -> StatsWdiget
