+++ ui/src/lib/components/StatsWidget/IconWidget.svelte
