+++ ui/src/lib/components/StatsWidget/ProgressBarWidget.svelte
