+++ ui/src/lib/components/StatsWidget/component.svelte
