+++ ui/src/lib/components/StatsWidget/types.ts
