+++ ui/src/lib/components/index.ts
@@ -2,12 +2,13 @@
 // SPDX-FileCopyrightText: 2024-Present The UDS Authors
 
 export { default as AnsiDisplay } from './AnsiDisplay/component.svelte'
-export { default as Link } from './Link/component.svelte'
-export { default as ProgressBar } from './ProgressBar/component.svelte'
-export { default as StatWidget } from './StatWidget/component.svelte'
-export { default as Tooltip } from './Tooltip/component.svelte'
 export { default as Card } from './k8s/Card/component.svelte'
 export { default as DataTable } from './k8s/DataTable/component.svelte'
 export { default as Drawer } from './k8s/Drawer/component.svelte'
-export { default as EventList } from './k8s/Event/EventList.svelte'
 export { default as Event } from './k8s/Event/component.svelte'
+export { default as EventList } from './k8s/Event/EventList.svelte'
+export { default as Link } from './Link/component.svelte'
+export { default as ProgressBar } from './ProgressBar/component.svelte'
+export { default as StatsWidget } from './StatsWidget/component.svelte'
+export { default as Tooltip } from './Tooltip/component.svelte'
+
